Move boards code under app/products/boards (mirroring playbooks layout)
BoardView is a boards-specific feature, so its model, schema, transformer,
and handler should live alongside the rest of the product code rather than
in the top-level database namespace. Adopt the same layout used by
playbooks and agents.

Structure:
  app/products/boards/
    constants/database.ts                       (BOARDS_TABLES)
    database/
      models/{index.ts,board_view.ts}
      schema/{index.ts,board_view.ts}
      operators/
        transformers/{index.ts,index.test.ts}
        handlers/{index.ts,index.test.ts}
    types/
      api.d.ts                                   (ambient BoardView/KanbanProps/...)
      database/models/board_view.ts              (BoardViewModel interface)

Wiring:
- tsconfig.json + babel.config.js: add @boards/* alias
- DatabaseManager (+ mock): import BoardViewModel from @boards
- Server schema index: import BoardViewSchema from @boards
- ServerDataOperator: import BoardsHandler from @boards
- Migration: import BOARD_VIEW from @boards/constants/database
- MM_TABLES.SERVER no longer carries BOARD_VIEW
- HandleBoardViewsArgs moved into the boards handler module
  (matches the playbooks pattern of co-locating handler args)

PropertyField/PropertyValue remain at the top-level operator because they
are shared infrastructure not tied to boards.
