+++ app/constants/database.ts
@@ -48,7 +48,6 @@ export const MM_TABLES = {
         THREAD_PARTICIPANT: 'ThreadParticipant',
         TEAM_THREADS_SYNC: 'TeamThreadsSync',
         USER: 'User',
-        BOARD_VIEW: 'BoardView',
     },
 };
