+++ app/database/migration/server/index.ts
@@ -23,11 +23,11 @@ const {
     PROPERTY_FIELD,
     PROPERTY_VALUE,
     SCHEDULED_POST,
-    BOARD_VIEW,
 } = MM_TABLES.SERVER;
 
 const {PLAYBOOK_RUN, PLAYBOOK_CHECKLIST, PLAYBOOK_CHECKLIST_ITEM, PLAYBOOK_RUN_ATTRIBUTE, PLAYBOOK_RUN_ATTRIBUTE_VALUE} = PLAYBOOK_TABLES;
 const {AI_BOT, AI_THREAD} = AGENTS_TABLES;
+const {BOARD_VIEW} = BOARDS_TABLES;
 
 export default schemaMigrations({migrations: [
     {
