+++ app/database/operator/server_data_operator/index.ts
@@ -2,8 +2,8 @@
 // See LICENSE.txt for license information.
 
 import AgentsHandler, {type AgentsHandlerMix} from '@agents/database/operators/handlers';
+import BoardsHandler, {type BoardsHandlerMix} from '@boards/database/operators/handlers';
 import ServerDataOperatorBase from '@database/operator/server_data_operator/handlers';
-import BoardsHandler, {type BoardsHandlerMix} from '@database/operator/server_data_operator/handlers/boards';
 import CategoryHandler, {type CategoryHandlerMix} from '@database/operator/server_data_operator/handlers/category';
 import ChannelHandler, {type ChannelHandlerMix} from '@database/operator/server_data_operator/handlers/channel';
 import CustomProfileHandler, {type CustomProfileHandlerMix} from '@database/operator/server_data_operator/handlers/custom_profile';
