+++ app/products/boards/constants/database.ts
@@ -1,0 +1,6 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+export const BOARDS_TABLES = {
+    BOARD_VIEW: 'BoardView',
+};
