+++ babel.config.js
@@ -18,6 +18,7 @@ module.exports = {
                 '@actions': './app/actions',
                 '@agents': './app/products/agents',
                 '@assets': './dist/assets/',
+                '@boards': './app/products/boards',
                 '@calls': './app/products/calls',
                 '@client': './app/client',
                 '@components': './app/components',
