+++ tsconfig.json
@@ -42,6 +42,7 @@
             "@actions/*": ["app/actions/*"],
             "@agents/*": ["app/products/agents/*"],
             "@assets/*": ["dist/assets/*"],
+            "@boards/*": ["app/products/boards/*"],
             "@calls/*": ["app/products/calls/*"],
             "@client/*": ["app/client/*"],
             "@components/*": ["app/components/*"],
